Fix implementation of get-id, add pure version
Before this commit, the implementation of twitter-rest/get-id was broken
-- if passed a Status object (which is intended to be possible), it
would return the ID of the status, not the ID of the user.

This commit adds a separate pure (and correct) version of the function that does not
fall back on the API, as well as tests for same. The twitter-rest
version wraps that, but falls back on the API when necessary.