Add a TwitterAPI protocol for indirection
This will make it much, much easier to mock out interactions with the
Twitter API -- a necessary prerequisite for storing and replaying
classifications.

The change is intended to be backwards compatible for now. If no object
is provided, it just creates one from the bound credentials. This
involves some overhead, but probably not a lot.